Move write file functions to a new file
